requirements.txt
# AI Agent Dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1

# LangChain
langchain==0.1.9
langchain-openai==0.0.6
langchain-community==0.0.24

# Data Processing
pandas==2.2.0
openpyxl==3.1.2

# Vector Store
chromadb==0.4.22

# Frontend
streamlit==1.31.1

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
pydantic==2.6.1
python-multipart==0.0.9
